heavily refactor ancestry
